แก้ไข design

+++ src/pages/admin/student/StudentDetail.js
@@ -31,26 +31,48 @@ function StudentDetail() {
     })
 
     const deleteStudent = (userID) => {
-        axios.delete(process.env.REACT_APP_API_URL + "/student", { data: { userID: userID } })
-            .then((response) => {
-                setStudentList(
-                    studentlist.filter((_) => {
-                        return _.userID !== userID;
-                    })
-                )
+        Swal.fire({
+            title: 'ต้องการลบนิสิตหรือไม่?',
+            showDenyButton: true,
+            showCancelButton: true,
+            confirmButtonText: 'ใช่',
+            denyButtonText: `ไม่ใช่`,
+            cancelButtonText: 'ยกเลิก'
+        })
+            .then((results) => {
+                if (results.isConfirmed) {
+                    axios.delete(process.env.REACT_APP_API_URL + "/student", { data: { userID: userID } })
+                        .then((response) => {
+                            setStudentList(
+                                studentlist.filter((_) => {
+                                    return _.userID !== userID;
+                                })
+                            )
 
-                Toast.fire({
-                    icon: 'success',
-                    title: 'Delete data success'
-                })
-                    .then(() => { window.location.href = "/admin/home"; })
+                            // Toast.fire({
+                            //     icon: 'success',
+                            //     title: 'Delete data success'
+                            // })
+                            Swal.fire({
+                                // position: "top-end",
+                                icon: "success",
+                                title: "Delete data success",
+                                showConfirmButton: false,
+                                timer: 1000,
+                              })
+                                .then(() => { window.location.href = "/admin/home"; })
 
 
-            }).catch(function (error) {
-                if (error.response) {
-                    console.log(error.response);
+                        }).catch(function (error) {
+                            if (error.response) {
+                                console.log(error.response);
+                            }
+                        });
                 }
-            });
+                else if (results.isDenied){
+                    window.location.href = "/admin/student/detail/" + userID;
+                }
+            })
     }
 
     const fetchData = () => {
@@ -125,30 +147,37 @@ function StudentDetail() {
                         <div className=" font-bold text-4xl m-10 grid grid-cols-1 place-items-center">ข้อมูลนิสิต</div>
                         <div className=' flex flex-row-reverse  '>
                             <div className='   mr-3'>
-                                <button onClick={() => gotoStudentEdit(userID)} className="relative inline-flex items-center justify-center p-4 px-6 py-3 overflow-hidden font-medium text-black transition duration-300 ease-out border-2 border-orange-300 rounded-full shadow-md group">
-                                    <span className="absolute inset-0 flex items-center justify-center rotate-180 w-full h-full text-white duration-300 -translate-x-full bg-orange-300 group-hover:translate-x-0 ease">
+                                <button onClick={() => gotoStudentEdit(userID)} className="relative inline-flex items-center justify-center p-4 px-6 py-3 overflow-hidden font-medium text-black transition duration-300 ease-out border-2 border-orange-400 rounded-full shadow-md group">
+                                    <span className="absolute inset-0 flex items-center justify-center rotate-180 w-full h-full text-white duration-300 -translate-x-full bg-orange-400 group-hover:translate-x-0 ease">
                                         <svg className="w-6 h-6 rotate-180" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M14 5l7 7m0 0l-7 7m7-7H3"></path></svg>
                                     </span>
-                                    <span className="absolute flex items-center justify-center w-full h-full text-orange-300 transition-all duration-300 transform group-hover:translate-x-full ease">แก้ไข</span>
+                                    <span className="absolute flex items-center justify-center w-full h-full text-black transition-all duration-300 transform group-hover:translate-x-full ease">แก้ไข</span>
                                     <span className="relative invisible">Button Text</span>
                                 </button>
                             </div>
-
-                            <div className='  mr-3'>
-                                <button onClick={() => goToAddWorkHistoryList(userID)} className="relative inline-flex items-center justify-center p-4 px-6 py-3 overflow-hidden font-medium text-indigo-600 transition duration-300 ease-out border-2 border-orange-300 rounded-full shadow-md group">
-                                    <span className="absolute inset-0 flex items-center justify-center w-full h-full text-white duration-300 -translate-x-full bg-orange-300 group-hover:translate-x-0 ease">
+                            <div className='  mr-3 place-items-end grid'>
+                                <button onClick={() => goToWorkHistoryList(userID)} className="relative inline-flex items-center justify-center p-4 px-6 py-3 overflow-hidden font-medium text-indigo-600 transition duration-300 ease-out border-2 border-orange-400 rounded-full shadow-md group">
+                                    <span className="absolute inset-0 flex items-center justify-center w-full h-full text-white duration-300 -translate-x-full bg-orange-400 group-hover:translate-x-0 ease rotate-180">
+                                        <svg className="w-6 h-6 rotate-180" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M14 5l7 7m0 0l-7 7m7-7H3"></path></svg>
+                                    </span>
+                                    <span className="absolute flex items-center justify-center w-full h-full text-black transition-all duration-300 transform group-hover:translate-x-full ease">ประวัติการทำงาน</span>
+                                    <span className="relative invisible">Button Text</span>
+                                </button>
+                            </div>
+                            {/* <div className='  mr-3'>
+                                <button onClick={() => goToAddWorkHistoryList(userID)} className="relative inline-flex items-center justify-center p-4 px-6 py-3 overflow-hidden font-medium text-indigo-600 transition duration-300 ease-out border-2 border-orange-400 rounded-full shadow-md group">
+                                    <span className="absolute inset-0 flex items-center justify-center w-full h-full text-white duration-300 -translate-x-full bg-orange-400 group-hover:translate-x-0 ease">
                                         <svg className=' text-white' width="30" height="15" viewBox="0 0 30 30" fill="none" xmlns="http://www.w3.org/2000/svg">
                                             <path d="M2 15.22H14.72M14.72 15.22H27.44M14.72 15.22V2.5M14.72 15.22V27.94" stroke="currentColor" strokeWidth="3.18" strokeLinecap="round" strokeLinejoin="round" />
                                         </svg>
                                     </span>
-                                    <span className="absolute flex items-center justify-center w-full h-full text-orange-300 transition-all duration-300 transform group-hover:translate-x-full ease">เพิ่มประวัติการทำงาน</span>
+                                    <span className="absolute flex items-center justify-center w-full h-full text-black transition-all duration-300 transform group-hover:translate-x-full ease">เพิ่มประวัติการทำงาน</span>
                                     <span className="relative invisible">Button Text</span>
                                 </button>
-                            </div>
-
+                            </div> */}
                             <div className='  mr-3'>
-                                <button onClick={() => deleteStudent(userID)} className="relative inline-flex items-center justify-center p-4 px-6 py-3 overflow-hidden font-medium text-indigo-600 transition duration-300 ease-out border-2 border-orange-300 rounded-full shadow-md group">
-                                    <span className="absolute inset-0 flex items-center justify-center w-full h-full text-white duration-300 -translate-x-full bg-orange-300 group-hover:translate-x-0 ease">
+                                <button onClick={() => deleteStudent(userID)} className="relative inline-flex items-center justify-center p-4 px-6 py-3 overflow-hidden font-medium text-indigo-600 transition duration-300 ease-out border-2 border-red-500 rounded-full shadow-md group">
+                                    <span className="absolute inset-0 flex items-center justify-center w-full h-full text-white duration-300 -translate-x-full bg-red-500 group-hover:translate-x-0 ease">
                                         <svg width="20" className=' text-white' height="20" viewBox="0 0 47 51" fill="#FFFFFF" xmlns="http://www.w3.org/2000/svg">
                                             <path d="M39.2592 23.4346V46.2701C39.2592 47.0752 38.6673 47.7277 37.937 47.7277H9.72969C8.99945 47.7277 8.40747 47.0752 8.40747 46.2701V23.4346" stroke="black" strokeWidth="6.54545" strokeLinecap="round" strokeLinejoin="round" />
                                             <path d="M19.4258 38.0104V23.4346" stroke="black" strokeWidth="6.54545" strokeLinecap="round" strokeLinejoin="round" />
@@ -156,7 +185,7 @@ function StudentDetail() {
                                             <path d="M43.6665 13.7172H32.648M32.648 13.7172V5.45759C32.648 4.65259 32.0561 4 31.3258 4H16.3407C15.6105 4 15.0185 4.65259 15.0185 5.45759V13.7172M32.648 13.7172H15.0185M4 13.7172H15.0185" stroke="black" strokeWidth="6.54545" strokeLinecap="round" strokeLinejoin="round" />
                                         </svg>
                                     </span>
-                                    <span className="absolute flex items-center justify-center w-full h-full text-orange-300 transition-all duration-300 transform group-hover:translate-x-full ease">ลบ</span>
+                                    <span className="absolute flex items-center justify-center w-full h-full text-black transition-all duration-300 transform group-hover:translate-x-full ease">ลบ</span>
                                     <span className="relative invisible">Button Text</span>
                                 </button>
                             </div>
@@ -187,7 +216,7 @@ function StudentDetail() {
                                     </> : <></>
                             }
                             <div className=" grid grid-cols-1 place-items-center">
-                                <div className=" block bg-gray-200 w-11/12 p-auto rounded-2xl">
+                                <div className=" block bg-gray-200 w-11/12 p-auto rounded-2xl ring ring-black">
                                     <div className=" flex justify-around">
                                         <div className=" ml-7">
 
@@ -316,16 +345,6 @@ function StudentDetail() {
                             </div>
                         </div>
                     </div>
-                    <div className='  mr-3 place-items-end grid'>
-                        <button onClick={() => goToWorkHistoryList(userID)} className="relative inline-flex items-center justify-center p-4 px-6 py-3 overflow-hidden font-medium text-indigo-600 transition duration-300 ease-out border-2 border-orange-300 rounded-full shadow-md group">
-                            <span className="absolute inset-0 flex items-center justify-center w-full h-full text-white duration-300 -translate-x-full bg-orange-300 group-hover:translate-x-0 ease rotate-180">
-                                <svg className="w-6 h-6 rotate-180" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M14 5l7 7m0 0l-7 7m7-7H3"></path></svg>
-                            </span>
-                            <span className="absolute flex items-center justify-center w-full h-full text-orange-300 transition-all duration-300 transform group-hover:translate-x-full ease">ประวัติการทำงาน</span>
-                            <span className="relative invisible">Button Text</span>
-                        </button>
-                    </div>
-
                 </div>
             )}
